pull rich text into just one file
