+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "textangular",
-  "version": "1.5.2-patreon",
+  "version": "1.5.2-patreon-rc1",
   "description": "A radically powerful Text-Editor/Wysiwyg editor for Angular.js",
   "maintainers": [
     {
