--- src/patreon_rich_text/index.js
@@ -1,7 +1,0 @@
-import angular from 'angular';
-import PatreonRichText from './patreon-rich-text';
-
-export const angularModuleName = 'PatreonRichText';
-
-angular.module(angularModuleName, [])
-    .factory('PatreonRichText', PatreonRichText);
